// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck
/*
 * This file exports all enum related types from the schema.
 *
 * 🟢 You can import this file directly.
 */

export const Role = {
	NEEDY: "NEEDY",
	DONOR: "DONOR",
	ADMIN: "ADMIN",
} as const;

export type Role = (typeof Role)[keyof typeof Role];

export const AuthProvider = {
	CREDENTIAL: "CREDENTIAL",
	GOOGLE: "GOOGLE",
} as const;

export type AuthProvider = (typeof AuthProvider)[keyof typeof AuthProvider];

export const UserStatus = {
	ACTIVE: "ACTIVE",
	BLOCKED: "BLOCKED",
	DELETED: "DELETED",
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const DonationRequestStatus = {
	PENDING: "PENDING",
	VERIFIED: "VERIFIED",
	REJECTED: "REJECTED",
	COMPLETED: "COMPLETED",
} as const;

export type DonationRequestStatus =
	(typeof DonationRequestStatus)[keyof typeof DonationRequestStatus];

export const DonationStatus = {
	PENDING: "PENDING",
	COMPLETED: "COMPLETED",
	CANCELLED: "CANCELLED",
} as const;

export type DonationStatus =
	(typeof DonationStatus)[keyof typeof DonationStatus];